Refactor handshake
